Revert "Add msg codec (#5)" (#6)
This reverts commit e0fc7aaa44005d8f28b469409905c09a7c9d0c92.

+++ src/common/connection.rs
@@ -1,5 +1,5 @@
 #[derive(Clone)]
 pub struct Conection {
-    pub id: u128,
+    pub id: u32,
     pub ip: String,
 }
